steps: Move mock server related steps into their own module
